Fix MapRDD join and add an example
The last version of the join was doing it wrong. It only did a join
between each silos (s_1...s_n with t_1...t_n, as a cartesian product),
but stopped there instead of returning a single silo per join (so s1
with all the t). A merge phase was added to solve that.